fix: update data types for time calculations and formatting in ResultScreen and FormatHelper

+++ lib/utils/format_helper.dart
@@ -1,5 +1,5 @@
 class FormatHelper {
-  static String formatTimeToMinutesSeconds(double totalSeconds) {
+  static String formatTimeToMinutesSeconds(int totalSeconds) {
     final minutes = (totalSeconds / 60).floor();
     final seconds = (totalSeconds % 60).round();
